make IsSecret and IsManagedPolicy fields public
