+++ internal/pkg/addons/addons_test.go
@@ -5,8 +5,6 @@ package addons
 
 import (
 	"bytes"
-	"io/ioutil"
-	"path/filepath"
 	"testing"
 
 	"github.com/aws/amazon-ecs-cli-v2/internal/pkg/addons/mocks"
@@ -70,54 +68,6 @@ func TestAddons_Template(t *testing.T) {
 			// THEN
 			require.Equal(t, tc.wantedErr, gotErr)
 			require.Equal(t, tc.wantedTemplate, gotTemplate)
-		})
-	}
-}
-
-func TestOutputs(t *testing.T) {
-	testCases := map[string]struct {
-		testdataFileName string
-
-		wantedOut []Output
-		wantedErr error
-	}{
-		"parses valid CFN template": {
-			testdataFileName: "template.yml",
-
-			wantedOut: []Output{
-				{
-					Name:            "AdditionalResourcesPolicyArn",
-					isManagedPolicy: true,
-				},
-				{
-					Name:     "MyRDSInstanceRotationSecretArn",
-					isSecret: true,
-				},
-				{
-					Name: "MyDynamoDBTableName",
-				},
-				{
-					Name: "MyDynamoDBTableArn",
-				},
-				{
-					Name: "TestExport",
-				},
-			},
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			// GIVEN
-			template, err := ioutil.ReadFile(filepath.Join("testdata", tc.testdataFileName))
-			require.NoError(t, err)
-
-			// WHEN
-			out, err := Outputs(string(template))
-
-			// THEN
-			require.Equal(t, tc.wantedErr, err)
-			require.ElementsMatch(t, tc.wantedOut, out)
 		})
 	}
 }
